Agents/local_requirements.txt
ag-ui-protocol==0.1.15
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.91.0
anyio==4.13.0
APScheduler==3.11.2
argcomplete==3.6.3
attrs==26.1.0
Authlib==1.6.9
beartype==0.22.9
boto3==1.42.85
botocore==1.42.85
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cohere==5.21.1
cryptography==46.0.7
cyclopts==4.10.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
edge-tts==7.2.8
email-validator==2.3.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
fastavro==1.12.1
fastmcp==3.2.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
genai-prices==0.0.56
google-auth==2.49.1
google-genai==1.71.0
googleapis-common-protos==1.74.0
greenlet==3.3.2
griffelib==2.0.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jeepney==0.9.0
jiter==0.13.0
jmespath==1.1.0
jsonpath-python==1.1.5
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
logfire-api==4.31.0
markdown-it-py==4.0.0
mdurl==0.1.2
more-itertools==11.0.1
multidict==6.7.1
nexus-rpc==1.4.0
notebooklm-py==0.3.4
ollama==0.6.1
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==25.0
pathable==0.5.0
platformdirs==4.9.4
playwright==1.58.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
py-key-value-aio==0.4.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-evals==1.78.0
pydantic-graph==1.78.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.1
Pygments==2.20.0
PyJWT==2.12.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
python-telegram-bot==22.7
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
s3transfer==0.16.0
SecretStorage==3.5.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.4
starlette==1.0.0
tabulate==0.10.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
types-protobuf==7.34.1.20260408
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uncalled-for==0.3.1
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wrapt==1.17.3
yarl==1.23.0
yt-dlp==2026.3.17
zipp==3.23.0
